Create unit tests for AutoETS
Test AutoETS on airline (built-in) and gas (imported from R) datasets, with both default conditions and default + allow_multiplicative_trend=True. Discrepancies occur in damped values, potentially due to different ways of calculating the values of information criterion in the two methods (the values for damped and undamped cases were very close), and therefore is not considered while comparing models.