Moving my implementation to contrib directory
As crew members suggested, I fixed some errors and put the suggested classifier in the contrib/distance_based file. 

As a conclusion of my research, the main reason my implementation tends to be take less time than the original one is the use of the 'dtaidistance' package as a tool for calculating distance that makes it faster.