[BUG] 1523 fixing forecast horizon bug (#1830)
Removes use of `_get_freq` in the `fh.to_absolute()` conversion. `_get_freq` removes anchorings, this was causing issues when users needed to convert a frequency involving an anchoring as the information was lost and the horizon defaults back to "W-SUN" even if another day e.g. "W-WED" was specified.