[BUG] fixed state change caused by `ThetaForecaster.predict_quantiles` (#2108)
Contrary to specification `ThetaForecaster.predict_quantiles` changed object attributes - namely, an attribute `sigma_` was created in `self`. This was probably accidental.

The tests did not pick this up, since much of the generic test suite was deactivated for the new probabilistic forecasting methods.

This is being changed in https://github.com/alan-turing-institute/sktime/pull/2100, which picked up the issue.