[ENH] Transformer pipeline and dunder method (#2090)
This PR introduces a `TransformerPipeline` for use in pipeline building.
The `TransformerPipeline` inherits from `BaseTransformer` and `_HeterogeneousMetaEstimator`, and can take a list of transformers, or a list of `(str, transformer)` pairs. If no names are provided, names are created and then made unique.

The `TransformerPipeline` is automatically invoked when the `*` (i.e., `__mul__` and `__rmul__`) dunder methods are called from any transformer. If multiple `TransformerPipeline`s are multiplied, they are conveniently merged to a single `TransformerPipeline`, instead of creating a pipeline of pipelines.