[ENH] Classifier pipeline and dunder method (#2164)
This PR introduces an `sktime` compatible transformer/classifier pipeline, and extends the design study in #2090.
It also contains minor additions to the `_HeterogenousMetaEstimator`, and common pipeline related utility functions moved there.

The `ClassifierPipeline` inherits from `BaseClassifier` and `_HeterogeneousMetaEstimator`, and takes a list of transformers, or string-transformer pairs, and a classifier. Internally, it uses the `TransformerPipeline`

The `ClassifierPipeline` and `BaseClassifier` now implement the `rmul` right multiplication dunder method, which implements the "obvious" functionality of calling `ClassifierPipeline` and/or `TransformerPipeline` concatenation whenever a `BaseTransformer` child is encountered as `other`.

This allows easy creation of transformer/classifier pipelines by commands such as
```python
my_trafo1 * my_trafo2 * my_clf
```
where dispatch and class selection happens "under the hood" of the dunder methods.