[ENH] sklearn typing util (#2208)
This PR contains the following typing utilities related to sklearn:
* a function `is_sklearn_estimator` which returns True/False whether the argument is an `sklearn` estimator.
* a function `sklearn_scitype` which returns a string (`"classifier"`, `"regressor"` etc) denoting the scitype of the estimator.
* four shorthances `is_sklearn_classifier`, for `is_sklearn_estimator` logical-and `sklearn_scitype` equal to certain scitype.

This can be used in type checking, as `sklearn_scitype` is closer to the use of `type`, and avoids `sklearn` imports of mixin classes.

Note: the logic is intentionally different from `sklearn.base.is_classifier` etc, as that assumes certain attributes to be set. The functions above do *not* assume that the argument is an `sklearn` estimator, and they check sub-class membership only.