[ENH] Upgrade of `BaseTransformer` to use vectorization utility, hierarchical mtype compatibility (#2219)
This PR upgrades `BaseTransformer` to use the `VectorizedDF` utility (that `BaseForecaster` uses) instead of bespoke vectorization.
As a side benefit, this upgrades `BaseTransformer` to be compatible with `Hierarchical` inputs.

The PR also adds a `test_base` test suite which targets testing of the boilerplate functionality in `BaseTransformer`. Currently, this uses actual transformers for the test, but these could be replaced by a mock transformer.