[BUG] Fix windows bug with index freq in VetorizedDF.__getitem__ (#2279)
Fixes #2194. Tests for transformers on hierarchical data were failing for Windows because the series obtained when looping through `VetorizedDF` have frequency `None` in Windows and `"D"` for both Linux and Mac. The fix is to add a new function that uses `pd.infer_freq` to update the frequency of each series in `VetorizedDF.__getitem__`.