[BUG] FeatureUnion output column names fixed (#2324)
`FeatureUnion` was returning incorrect column labels due to a bad `str.join` command. This is now fixed.

The tests did not catch this since the format was correct, but not the number of columns.

This has been corrected, and a test has been added to ensure that `FeatureUnion` produces the right column names.