[ENH] Aggregator as a transformer (#2292)
This PR adds a aggregator transformer, which takes observations at the bottom level of the hierarchy and sums them according to the index provided. This is linked to PR #2287 and issue #2157

A transformer that takes hierarchical observations at the bottom level of the hierarchy only, and sums them appropriately using the index. An identifier is placed in the index for aggregated nodes called `__total`. The reconciler implemented in #2287 use this identifier, so probably this is a better one to focus on first.