[ENH] Implement reconcilers for hierarchical predictions - transformers (#2287)
This PR adds a reconciler transformer, which takes hierarchical predictions and processes them so they sum appropriately. Three methods are implemented here out of the list in issue #2157
 - bottom up 
 - ols
 - wls_var

Hierarchical reconciler which depends only on the structure of the hierarchy. The three methods are detailed [here](https://otexts.com/fpp3/hierarchical.html) and the code is constructed similar to the methodology as noted in these [slides](https://github.com/robjhyndman/reconciliation_review_talk/blob/main/10years_reconciliation.pdf). The only difference between the three methods is the construction of the "G" matrix.