[BUG] forecasting pipeline dunder fix (#2431)
This fixes a semi-bug in the dunders for forecasting pipelines.

Since forecasting pipelines were both forecasters and transformers, the transformer chain logic would take precedence as parsing of the expressions was currently coded, resulting in a transformer and not a forecaster when chaining transformers and a forecaster. Strictly speaking that was not wrong, since forecasting pipelines can be used as transformers.

This PR changes precedence so that the forecaster type takes precedence over the transformer type, ensuring that multiplying transformers with forecaster results in forecasters.