[ENH] remove error message on exogeneous X from DirRec reducer (#2463)
The `_DirRecReducer` did not comply with the generic forecaster interface, where exogeneous `X` should not raise an error (in the worst case it should be ignored). The reducer did raise an error.

This PR does not add use of exogeneous `X`, it simply makes the reducer interface compliant, by ignoring `X` instead of raising an error.