[ENH] more forecaster scenarios for testing: using `X` (#2462)
This PR enables one more forecasting scenario for testing, a scenario which passes exogeneous `X`. In return, the same scenario without `X` is disabled to save compute time.
The PR also renames the `pre-refactor` tag for scenarios into `is_enabled`, and also adds that tag to the one new scenario with `X`.