[ENH] cleaning up `_panel._convert` module and `loading_data` tutorial (#2519)
This PR does some minor cleaning in the `_panel._convert` module which had much copy-pasted legacy code:

* removal of commented out code
* removal of heavy input checks (this should be done by the `check` functions if needed)
* added module docstring

It also removes redundant and outdated parts in the `loading_data` tutorial, which was importing a lot of now private functionality of the `datatypes` module. This is replaced by references to the datatypes tutorial which explains the multiple data formats.