[ENH] Logger update - `__init__` removal, private `log` attribute (#2533)
This PR makes the following changes to `_MockEstimatorMixin`:

* changes the name of the log attribute from `_log` to `__log`. That way, the attribute becomes hidden and cannot be affected directly.
* write logic is slightly changed, with `log` getting a setter `add_log_item` that appends to the log (instead of overwritin it); `__init__` is removed entirely, first access creates the list and not the `__init__`.

Also adds the `utils` module to the ignore-list of the `all_estimators` retrieval utility, since (a) the test estimators are not actually ones a user will want to see, and (b) the combined mock/logger estimators change state in `predict` etc and fail some tests. But, for logger estimators a change of state in `predict` etc is actually intended (addition of the log).