[ENH] Differencer NA handling - "fill zero" parameter (#2487)
This PR solves #2478, it upgrades `Differencer` with NA handling parameter `na_handling`, which allows, besides the two current behaviours, a third behaviour `fill_zero` that fills zero.

This parameter replaces the boolean `drop_na` parameter, and adds deprecation with replacement at 0.13.0, and move to a `fill_zero` default.

Why is the `fill_zero` parameter a better default? It results in a "same index" and "no NA" condition at the end, which is the most compatible out of the alternatives when used in a pipeline. See also discussion (and user frustration) here: https://github.com/alan-turing-institute/sktime/issues/2452 and here: https://github.com/alan-turing-institute/sktime/issues/2451