[ENH] Removed unused interval_width parameter of Prophet (#2630)
Resolves: https://github.com/alan-turing-institute/sktime/issues/2626

To restate what is mentioned in the issue, the interval_width parameter of Prophet is not used and is later overwritten, and is not mentioned in the docstring. The initialization logic can be replaced by the default value if needed. So interval_width is removed as an __init__ argument and deleted where self.interval_width = interval_width.