[BUG] Forecasting pipeline get/set params fixed for dunder generated pipelines (#2619)
This fixes an unreported bug with `get_params` and `set_params` of the forecasting pipelines, where either method would break, since the inner versions referred to `steps` and not to `_steps` (like in `TransformerPipeline`, which was added).

Also adds a test that would catch this in the future - a test instance of either pipeline which is the result of dunder construction. Apparently this was accidentally omitted and not tested so far.