test_update_predict _predicted_index no gaps in data between fit and update (#2701)
Fixes the problem in test_update_predict _predicted_index of gaps between data passed in fit and update by replacing SlidingWindowSplitter with ExpandingWindowSlicer as described in #2700 + made the appropriate changes in the corresponding util function.

The problem with the test was that the first non-trivial update call uses data that are more than one time step ahead than the data passed in fit. As a result when we try to reconcile y_train and the subset of y_test from the splitter the resulting time series has gaps and frequency None. For some reason this was not causing any forecasting algorithm to break apart from BaggingForecaster in #2248 and in only 2 test cases.