[BUG] Fixed fit behaviour of `Imputer`, improved docstrings (#2382)
This PR addresses the concern raised in https://github.com/alan-turing-institute/sktime/discussions/2267 about information leakage on the test set when using the `Imputer` transformer in a forecasting pipeline.

Changes in this PR:
* improved docstrings, some corrections, and a highlight which methods are prone to information leakage
* changes `mean` and `median` methods to use the mean/median from fit
* changes `get_test_params` so all admissible method strings are tested. This should be a "fast" estimator so no huge overhead in terms of compute (hopefully)
* replaced `Z` by `X` in `_transform` and named the output `Xt`, for better readability throughout

Also fixes a bug in the example in `ForecastingGridSearchCV` which used an unknown method `"last"` for the imputer. This never raised an error, since the data in the example did not have nan in it (and error handling was less stringent than in this PR).