[ENH] Adding more tests for probabilistic metrics (#2683)
Implements testing for all cases of input to the probabilistic tests. Also fixes some issues that were discovered whilst testing. Namely fringe cases of having 0.5 quantiles given to an interval metric, this creates an interval with 0 coverage which causes some failures in the tests so I now remove this column inside the tests with a warning.