[ENH] Add MultiplexTransformer (#2738)
Implements a `MultiplexTransformer`. Related to #2670 and #2459.  Note this implementation is meant to be very similar to how `MultiplexForecaster` currently works - see #2458 making use of the `_DelegatedTransformer` in #2612.

With `MultiplexTransformer` one now in theory has a new way to multiplex over multiple transformers.  (Note one could optionally include various transformers with the pre-existing `OptionalPassthrough` but `MultiplexTransformer` should be useful when people want exactly one transformer.