[ENH] interface compliance test to ensure sklearn compliance of constructor (#2732)
This PR makes the following changes:
* reactivates the `check_constructor` test, which was not executed since the test refactor since we forgot to rename it to start with `test`
*  adds a test to `TestAllEstimators` which checks for interface non-compliance along the lines of the bug #2730, i.e., presence of a `**kwarg` signature element in `Class.__init__`.

Why does having `kwargs` signature elements mean interface non-compliance? `sktime` aims to be compliant with `scikit-learn`, and in particular the nested parameter interface of `scikit-learn`. That assumes that all parameters in the class constructors are explicitly named keyword arguments, and parameters exposed by the `scikit-learn` parameter interface are precisely the named keyword arguments. Passing parameters via unnamed keyword arguments is not caught by the parameter inspection and setter interface of `scikit-learn` or `sktime`. So far, this has been possible without setting off a test, and could lead to unexpected and buggy behaviour, as #2730 shows.