[ENH] Refactor estimator method iteration in `TestAllEstimators` as test fixture (#2781)
This PR refactors the "non state changing method" that is looped over in many tests, in `TestAllEstimators`, as a string-valued fixture.

This PR builds on https://github.com/alan-turing-institute/sktime/pull/2779 which introduces a conditional escape of `predict_proba` method. The refactor additional in this PR moves the multiple times copy-pasted copy-paste added by #2779 to one single location.