[DOC] cleaning up transformer API reference (#2818)
This PR cleans up the transformer API reference:

* a lot of transformers were simply missing. Included a number of transformers contributed by non-core-devs, and hierarchical transformers (aggregation, reconciliation). They were added.
* the high-level taxonomy of transformers was re-ordered, e.g., to be clear about what the return is (primitives, series, panel, etc). This is explained in a table at the start.
* new mid-level taxonomy of transformers. Parts correspond to the TSC estimator taxonomy, parts based on methodological details.

Further, it seems that the transformers have been placed in the `series` and `panel` folders more or less at random, historically. The taxonomy does not reflect well the folder structure, so we may like to move around things in the folders at some point in the future.