[ENH] Lagging transformer (#2783)
This PR introduces a transformer for lagging for use in pipelines and and potential component in more complex algorithms that rely on lagging, e.g., a reduction algorithm refactor. Implements #2526.

A particular challenge is to cover both range/integer and time-like indices, since `pandas.shift` handles those inconsistently.
In the transformer, this is handled by a manual step-out for integer indices.