[BUG] transformers: output type check fix for ambiguous return types (#2843)
Fixes #2841.

The output type checks in transformers could run into problems in the specific case where the inner `_transform` produced a `pd.DataFrame` that was compliant both with a `Panel` and a `Table` specification.

In cases where this happened, the transformer would error out.

This is easily fixed, since we *know*  when to expect a `Panel` or a `Table`, based on the output scitype. The bug has been fixed by using that tag to restrict the expected outputs, which prevents multiple type identifications.