[BUG] fix for `Differencer.inverse_transform` not having access to data index `freq` (#3007)
This fixes `Differencer`'s inverse transform not having access to the `freq` attribute, following deprecations of `freq` in various places in `pandas`. Relies on the recent upgrade of `freq` handling in `ForecastingHorizon` to future `pandas` compatibility with @khrapovs.

Fixes #2806 (and, by extension, #2879).
Contains a compressed version of the code in #2806, as test.