[MNT] remove custom `__repr__` from `BaseTask`, inherit from `BaseObject` (#3049)
This PR addresses imminent removal/change in `sklearn` `_pprint` and the private import in `sktime` `BaseTask`, and partially addresses #3037.

`BaseObject` has its own `__repr__` and could be used here, but `BaseTask` did not inherit from `BaseObject`.

To address the change in `__repr__`, `BaseTask` was changed to inherit from `BaseObject`, with some compatibility changes towards the `sklearn` interface. The complete removal of the custom `__repr__` in favour of `BaseObject.__repr__` should address this instance of `_pprint` import.