[ENH] ClearSky transformer for solar time series (#3130)
This PR is a new transformer which is intended to be used with solar (related) time-series. The method approximates the clear sky value for a given day-of-year and time-of-day based on the input time-series. Values are then transformed into a percentage domain using this estimate of the clear sky as the denominator.

The clear sky is approximated via a weighted quantile, where the weights are defined using von-mises kernels with bandwidths chosen by the user.

This transformation can be inaccurate at low values, in the solar example during early morning and late evening. Therefore, clear sky values below a threshold can be fixed to zero in the transformed domain. Denominator values of zero are set to zero in the transformed domain by default.

This transformer is based on the work detailed in this [paper](https://doi.org/10.1016/j.solener.2009.05.016)

Also adds a added a `load_solar()` data loader function which queries [this](https://www.solar.sheffield.ac.uk/pvlive/) dataset for the example