[BUG] Fix `ignores-exogeneous-X` tag in `DirectReductionForecaster` (#3257)
`DirectReductionForecaster` is meant to allow for exogenous variables. However, because it inherits from `BaseForecaster` the default tag for exogenous variables was: ` "ignores-exogeneous-X": True`. This meant that any instance of `DirectReductionForecaster` was ignoring any exogenous variables passed to it during training and prediction. This PR fixes the tag so that `DirectReductionForecaster` uses `X`.