[ENH] in `Lag`, make column naming consistent between single-lag and multi-lag case (#3261)
Problem:
`Lag` outputs column names prefixed with "lag_x__" when using multiple lags. If a single lag is specified then the output column name is the same as the original column name (if the input is a `pd.DataFrame` rather than a `pd.Series`) without "lag_x__" prefixed to the original column name. It would be desirable for the output column names to be consistent, irrespective if a single lag or multiple lags is specified.

Fix implemented:
`Lag` now outputs the same column name irrespective of whether a single lag or multiple lags is specified.