[ENH] test `pd.Series` with name attribute in forecasters (#3323)
This changes one of the univariate `pd.Series` `y` fixtures in the scenarios for forecasters to have a `name` attribute.

Relies on https://github.com/alan-turing-institute/sktime/pull/3290. Combined with that, this change additionally tests that the `name` is preserved and is handled without exceptions.

Copy of reverted #3297, which made the change to an unused fixture, mistakenly. This has been fixed, the change is now in a fixture that is used.