[BUG] changed wrong docstring default value of `start_with_window` in `SlidingWindowSplitter` to actual default value (#3340)
Fixes #3315.
Changes wrong docstring default value of `start_with_window` in `SlidingWindowSplitter` to actual default value.

The default value was mentioned to be False, but it should be True, in the code, it has been rightly used but, in the docstring, there was an error. Changed the default value in the docstring.