[ENH] global/local setting for `DirectReductionForecaster` (#3327)
This introduces a parameter `pooling` to `DirectReductionForecaster`, which can be set to `"global"` for global pooling, `"local"` for per-instance pooling, and `"panel"` for panel-wise pooling. Pooling refers to the `X` and `y` passed to the supervised sklearn regressors, which could be constructed from single instances, or the entire panel/hierarchy, in case there are more than one instances.

Giving the user a choice is the way to go here imo, as all settings can make sense.

The logic of `DirectReductionForecaster` also had to be extended to also accommodate for potential multi-indexed `pandas.DataFrame` appearing in the `fit`/`predict` parts. Unfortunately this is a bit fiddly, since it seems `pandas` interfaces are quite discrepant between the 1-level case and the 2- and multi-level case.