[ENH] `BaseTransformer` data memory - enabled by tag (#3307)
Variant of https://github.com/alan-turing-institute/sktime/pull/3306 which introduces a `remember_data` tag. Data is only remembered and stored to `self._X` if the `remember_data` tag is set.

This PR adds a data memory to transformers, similar to forecasters.

If enabled by the tag, all data seen are stored in an attribute `self._X`, as a reference if possible (due to coercions, storing a reference may not be possible).

This allows to refactor some transformers which already to that internally, and remove a layer of boilerplate:

* `Imputer`
* `Lag`
* `IndexSubset`
* `PandasTransformAdaptor`
* `WindowSummarizer`

The tag is set to `False` by default, and to `True` only in estimators that currently use `self._X` already (the refactored transformers listed in #3306).

Because `self._X` may now have to be written, we need to check that `fit_is_empty` is `False` whenever `remember_data` is `True`. This is checked by an added transformer test.