[ENH] Fourier features transformer (#3374)
Implements Fourier features transformer, #2650. See also #1694.

This transformer is applied on the time index to generate Fourier series features as described here -> https://robjhyndman.com/hyndsight/longseasonality/

When fit is called the minimum date of X is stored. This date serves as the reference point for the time-steps passed, t.
So in the transform step, t is going to be the difference between the minimum date of X passed on fit (based on the frequency of the period index). 

This behaviour is checked in the test `sktime.transformers.series.tests.test_fourier.test_fit_transform_behaviour`