[ENH] signature kernel from (Kiraly et al, 2016) (#3355)
Adds the signature kernel from (Kiraly et al, 2016) and implements #1085.

Contains a modern kernel/pairwise transformer interface, and some historical legacy code to allow easy reproduction of the experiments reported in the 2016 paper.

This could be refactored more nicely, e.g., turning the low rank representation into an object with methods add/mult (not just collection). but this is simply an 1:1 adaptation into the `sktime` interface with linting, docstrings, and minor cleanup.

Fun fact, I was apparently already using the `numpy3D` mtype when I wrote this code in 2015.