[ENH] second param sets for selected estimators (#3428)
Towards #3429. This adds a second parameter set for all estimators checked via `check_estimator` in the `no-softdeps` CI element.

This is generally useful, and also allows https://github.com/alan-turing-institute/sktime/pull/2862 to pass that CI element.

Also fixes a bug discovered through this:
`ExponentTransformer.inverse_transform` breaking if `power` is close to zero. This is now dealt with by a skip and a warning.