[MNT] update `test_interpolate` to be `pandas 1.5.0` compatible (#3467)
This updates `test_interpolate` to be `pandas 1.5.0` compatible.

Apparently, it is no longer possible to assign `pd.Series` to fields of `pandas.DataFrame` via `iloc` or `at`.

Assignment via `iloc` breaks, and `at` now seems to result in `numpy` rather than `pd.Series`. Bug or not, this is fixed by accessing `values` directly.