[MNT] loosen `scipy` bound to <2.0.0 (#3587)
This PR loosens the upper version bound for `scipy` to `<2.0.0`, thus removing any restriction on `1.X.X` versions.

We had the MINOR bound introduced after repeated breakage, this was mostly on `sktime` side using private methods (which was a mistake). Afaik we are no longer doing that, and `scipy` is very good at deprecation, so I think we can safely move the bound up to the MAJOR cycle.

Fixes #3586.