[ENH] Save/Load Functionality for DL Estimators (#3425)
Fixes #3022 - adds save/load functionality for DL estimators.
Also changes the file format of file saves slightly, but this has not been released yet and therefore does not require deprecation.

The Save/Load function for classical estimators is from #3336 and has been expanded for DL models along the same design. Keras is known to be problematic for certain kinds of (de)serialization, so this uses `model.save()` from `keras` along with `pickle.dump()`. The estimators are saved as `.zip` files.

Components of this directory include:
1. Metadata
2. Object
3. Keras History
4. Keras Model subdirectory

The first two components are the same as previously mentioned PR, stored differently to [avoid](https://github.com/alan-turing-institute/sktime/pull/3336#issuecomment-1242013162) this problem.

All the attributes are saved and restored (optimizers/history etc.) appropriately.