[BUG] Fix `ConformalIntervals` update does not update residuals_matrix (#3460)
Fixes #3419 

As described in the bug issue, the `update` routine does not update the `residuals_matrix` which means the conformal intervals will be incorrect on the updated data. This PR computes the sliding residuals on the remaining part of the `residuals_matrix` in the case of an update.

This PR adds new rows and new columns populated for the new rows, but does not extend the existing rows as it stands. Is this an acceptable approximation? The way it is used this will only matter where old rows are not extending the length of the forecasting horizon. Happy to explain this aspect in more detail if this doesn't make sense.