[GOV] clarifications on algorithm maintainers (#3676)
This PR is a proposal following my action to clarify certain points in the governance document. 

As it was discussed in the Oct 25 CC meeting, the original #3154 was contentious in its changes to the algo maintainer role, this should be the part that was **contentious** to @aiwalter. The other part is in #3675. Relevant action here: https://github.com/sktime/community-org/issues/40

Changes are clarifications to the role of an algorithm maintainer:

  * the responsibility they have with respect to classes/estimators is clarified
  * it is clarified that the responsibility and ownership is to/of python classes, not abstract concepts
  * in particular, algorithm maintainers have no rights/responsibilities regarding other instances of the same abstract concept
  * algorithm maintainers are better delineated from core developers