[ENH] Add `colors` argument to `plot_series` (#3908)
Add an argument `colors` to `plot_series` to allow the user to pass a list of colors. 

Also adds a small util function to check that the length is correct, and use `from matplotlib.colors import is_color_like` to check that they are all valid colors. If either condition is not met, then it emits a warning and defaults back to the current behavior, which is using `sns.color_palette("colorblind", n_colors=n_series)`.