[BUG] in `ConformalIntervals`, fix update of residuals matrix for sliding window splitter (#3914)
Fixes https://github.com/sktime/sktime/issues/3913

Ensures only indices in the new residuals matrix are populated if already calculated and avoids the KeyError.

Also includes a test. While thinking about the test I realised a limitation of the `_update` method. It is set up to not recalculate `residuals_matrix` rows which have already been calculated. However, because the additional rows will add new columns, you end up with the existing rows having NULL in the last x columns. This has the potential to cause issues if for example the window cv is long and you are building a lot of the matrix via the `_update` method. 

I have fixed by defining an `_extend_residuals_matrix` function which uses the already fitted forecasters and extends the prediction to the length of the new `y_test`.

I have added a test to check `ConformalIntervals` and `NaiveVariance` with strategy = `refit` or `update` and `cv` =  sliding or expanding. The test threw up an erroneous assertion error in `PinballLoss` which I have also added a fix for (we use `convert_dtypes()` from pandas which converts to `Float64` instead of `float`).