[ENH] Check for frequency in hierarchical data, provide utility function to set frequency for hierarchical data (#3729)
Addresses concerns raised in https://github.com/sktime/sktime/issues/2710 (not strictly a bug).

RecursiveReducer will fail when frequency is not set, but setting and checking for frequency with multivariate data is tricky. 

I included a plausibility check in recursive reducer to give a plausible warning, and provided a convenience function set frequency (set_freq_hier). Open question: Should set_freq_hier be included and performed automatically? Currently not sure of all the places where to include it, reducers and temporal_train_split come to mind?