[BUG] Fix `StatsForecastAutoARIMA` preidct in sample incorrect start position (#3942)
In `StatsForecastAutoARIMA`, the forecast is always returned from the start time of the training, even though `fh` specifies to start at a moment in the middle of the training time range.

I think the reason is that the index used is incorrect, the `fh` passed in is not used correctly, and the past results of the prediction are always returned from the start time of the training. We should use `from_cutoff=True`. I did this and the final result was as expected.