[ENH] Update `_check_soft_dependencies` to allow PEP 440 specifier strings for version bounds (#3925)
This PR adds functionality to `_check_soft_dependencies` to parse and check PEP 440 package version specifier strings.

This is useful in a case where certain estimators require specific versions of dependencies, but this is not a requirement of the `sktime` framework itself.

An example where this could be useful is the case of version increments of `sktime` dependencies which do not break the framework but break individual estimators. The problems in those estimators would then be isolated to the estimators and do not prevent `sktime` version compatibility, i.e., of the framework code. Issues with individual estimators can then be resolved in a staggered way, not blocking framework releases for people who do not want to use that specific estimator (which is not an uncommon case as `sktime` has a number of rarely used ones, and those are usually also the ones that are less well maintained).

For a recent example, of this kind of failure, see the issues around the recent version increase of `sklearn` to 1.2.0:
https://github.com/sktime/sktime/pull/3922