[BUG] Fix `TimeSince` check of inconsistency between `time_index` and `start` (#4015)
In the `TimeSince` transformer, there is a check that the type of the time index and the `start` argument are consistent. The check was still being done on `X.index` which is fine for series data, but not for panel or hierarchical data. The check has been changed to the `time_index` variable.