[BUG] fix `ForecastX` predictions (#3987)
Fixes https://github.com/sktime/sktime/issues/3986 and includes consistency tests.

The reason for the bug is that columns may end up permuted, and the wrapped forecaster may or may not be able to cope with that. The fix is ensuring that the `X` in predict always has columns in the same order as in `fit`.

Probably, we should ensure that all estimators can deal with permuted columns by coercing in the base class, but that's a separate topic.