[ENH] Change `DateTimeFeatures` trafo to work with multi-index data and add option to drop columns. (#3996)
Changes the `transform` method in `DateTimeFeatures` to handle dataframes with a multi-index (e.g., panel and hierarchical data). The datetime features are now computed directly from the time index of the multi-index. This reduces the runtime significantly when using `DateTimeFeatures` with panel or hierarchical data; I measured a factor of 10 improvement.

A `keep_original_columns` argument is added which gives the option to keep or drop the columns in `X` passed to `.transform`. 

`transform` has also been refactored to avoid making a copy of `X`  to make it more memory efficient. 

See https://github.com/sktime/sktime/pull/3996#issuecomment-1364725220 for more context.