[BUG] fix `statsmodels` estimators when exogenous `X` is passed with more indices than `fh` (#3972)
Fixes [#3830](https://github.com/sktime/sktime/issues/3830)

Statsmodels with exogenous features would raise an exception in `predict` if `X` contained more indices than the range of `fh`.
This was due to all indices being passed through to `statsmodels` `predict`.  The fix ensures that only the expected `X` indices from the test set are passed to `statsmodels` `predict`.

Fixes #3830 as this bug was also the root cause of the exception raised when using `evaluate` with `statsmodels` based estimators.